instruction/requirements.txt
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
bitsandbytes>=0.43.3   
protobuf<4.0.0
hf_transfer>=0.1.8     
trl>=0.9.4
peft>=0.11.1
accelerate>=0.31.0     
flask>=3.0.3
nacos-sdk-python>=0.1.14
eth-utils>=4.1.1
xformers>=0.0.26.post1
loadtime==1.0.6
filelock==3.18.0